requirements.txt
# PDF extraction and text processing
PyMuPDF>=1.23.0
spacy>=3.7.0
tiktoken>=0.5.0
unidecode>=1.3.0

# OpenAI API (full pipeline + server)
openai>=1.0.0
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Metadata tracking
pandas>=2.0.0
openpyxl>=3.1.0

# Server
fastapi>=0.104.0
uvicorn>=0.24.0

# UI
tqdm>=4.66.0
rich>=13.7.0
